Isolate image processing logic
